feat: 아티클 박스 및 배지 컴포넌트 추가, 스타일 파일 생성
