+++ frontend/src/pages/article/ArticleBox/ArticleBox.styled.ts
@@ -1,0 +1,24 @@
+import styled from "@emotion/styled";
+
+export const ArticleContainer = styled.div`
+  display: flex;
+  flex-direction: column;
+  gap: 1.25rem;
+  padding-top: 2rem;
+`;
+
+export const ArticleHeader = styled.div`
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 0 1rem;
+`;
+
+export const ArticleIntro = styled.p`
+  color: #73798d;
+  font-size: 1.125rem;
+`;
+
+export const ArticleIntroText = styled.span`
+  color: #007bff;
+`;
