+++ frontend/src/pages/article/ArticleBox/CardList/Card/Badge/Badge.styled.ts
