+++ frontend/src/pages/article/ArticlePage.styled.ts
@@ -35,26 +35,3 @@ export const Box = styled.div`
   display: flex;
   justify-content: space-between;
 `;
-
-export const ArticleContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: 1.25rem;
-  padding-top: 2rem;
-`;
-
-export const ArticleHeader = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 0 1rem;
-`;
-
-export const ArticleIntro = styled.p`
-  color: #73798d;
-  font-size: 1.125rem;
-`;
-
-export const ArticleIntroText = styled.span`
-  color: #007bff;
-`;
