+++ frontend/src/pages/article/ArticlePage.tsx
@@ -2,21 +2,18 @@ import {
   ARTICLE_CATEGORY_ENTRY,
   type ArticleCategoryKey,
 } from "@domains/filter/articleCategory";
-import { ARTICLE_SORT_MAP } from "@domains/sort/article";
 import Tab from "@shared/components/Tab/Tab";
-import SortList from "../../domains/components/SortList/SortList";
+import ArticleBox from "./ArticleBox/ArticleBox";
 import * as S from "./ArticlePage.styled";
 import ArticleSearchBar from "./ArticleSearchBar/ArticleSearchBar";
-import CardList from "./CardList/CardList";
 import { useArticleCategory } from "./hooks/useArticleCategory";
 import useArticleList from "./hooks/useArticleList";
 import TagList from "./TagList/TagList";
 
-const DEFAULT_SORT_TYPE = "createdAt";
 const DEFAULT_ARTICLE_CATEGORY_TYPE = "all";
 
 function ArticlePage() {
-  const { refetch, isLoading, articles } = useArticleList();
+  const { refetch } = useArticleList();
   const { selectedCategory, updateCategory } = useArticleCategory(
     DEFAULT_ARTICLE_CATEGORY_TYPE,
   );
@@ -25,15 +22,6 @@ function ArticlePage() {
     key,
     label,
   }));
-
-  if (isLoading) return <div>Loading...</div>;
-  if (!articles) return <div>No articles found</div>;
-
-  const { articleContents, totalCount } = articles;
-
-  const handleSelect = () => {
-    refetch();
-  };
 
   const handleTabSelect = (key: ArticleCategoryKey) => {
     updateCategory(key);
@@ -57,20 +45,7 @@ function ArticlePage() {
         selected={selectedCategory}
       />
       <S.Box>
-        <S.ArticleContainer>
-          <S.ArticleHeader>
-            <S.ArticleIntro>
-              <S.ArticleIntroText>{totalCount}개</S.ArticleIntroText>의 아티클이
-              모여있어요.
-            </S.ArticleIntro>
-            <SortList<typeof ARTICLE_SORT_MAP>
-              sortMap={ARTICLE_SORT_MAP}
-              onSelect={handleSelect}
-              initialValue={DEFAULT_SORT_TYPE}
-            />
-          </S.ArticleHeader>
-          <CardList articles={articleContents} />
-        </S.ArticleContainer>
+        <ArticleBox />
         <TagList />
       </S.Box>
     </S.Main>
